+++ pkg/providers/cloudwrapper/cloudwrapper_errors.go
@@ -3,8 +3,8 @@ package cloudwrapper
 import "errors"
 
 var (
-	// ErrCloudWrapperProperties is returned when operation on cloudwrapper properties fails
-	ErrCloudWrapperProperties = errors.New("cloudwrapper properties")
 	// ErrCloudWrapperLocation is returned when operation on cloudwrapper location fails
 	ErrCloudWrapperLocation = errors.New("cloudwrapper location")
+	// ErrCloudWrapperLocations is returned when operation on cloudwrapper locations fails
+	ErrCloudWrapperLocations = errors.New("cloudwrapper locations")
 )