+++ pkg/providers/cloudwrapper/testdata/TestDataConfiguration/default.tf
@@ -1,0 +1,3 @@
+data "akamai_cloudwrapper_configuration" "test" {
+  id = 1
+}