+++ pkg/providers/cloudwrapper/testdata/TestDataConfiguration/no_config_id.tf
@@ -1,0 +1,1 @@
+data "akamai_cloudwrapper_configuration" "test" {}